pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "Research Article Summarization Assistant"
authors = [{ name = "OKHADIR Hamza", email = "hamza.okhadir2018@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "langchain-pinecone>=0.2.5",
    "langchain[groq]>=0.3.23",
    "python-dotenv>=1.1.0",
    "transformers>=4.51.3",
]
